Move all parser logic into parser-specific files. The index is now open to do more in the future if needed
